requirements.txt
faster-whisper~=0.6.0
torch~=2.0.1
websockets~=11.0.3
uvicorn~=0.20.0
fastapi~=0.96.0
PyAudio~=0.2.13
websocket-client~=1.5.1
rel~=0.4.9
sounddevice~=0.4.6
espeak-phonemizer>=1.1.0,<2
requests~=2.31.0
TTS~=0.15.0
gTTS~=2.3.2
playsound~=1.3.0
numpy~=1.22.0
onnxruntime~=1.12.0
python-multipart~=0.0.6